venvscript: added persistent options, fixed help
Fixed help message display. Now checks for amount of args before the rest of the checks to prevent errors.
Add persistent options --use-uv, --use-uvi, and --default-run. --use-uv forces the use of uv (installed outside venv) instead of pip. --use-uvi installs and forces the use of uv inside of the venv. --default-run causes the script to deafult to running the game instead of displaying a help message