Refactored venv script to use pathlib
The venv script used pathib a little bit before but never to its fullest potential. It now uses it for everything (except one thing). This commit also fixes an issue with pip introduced when I added support for pip. I was too lazy to separate it into its own commit (it doesn't matter either because this is my repo and I'm the only one making commits).