Add pathlib.Path support to isValidDirectory and add to loadPreferences
Adds support for pathlib.Path objects in isValidDirectory. Using them is 10 times slower but the checks should be a lot more accurate since pathlib does most of the magic itself.
Adds a check to loadPreferences to prevent loading of invalid values for the theme.
Changes a check in loadPreferences to be a lot faster.