[venv script] Rewrite config and add new arguements
Switch from an assortment of dot files to a single .cfg file for configuration.
Adds arguement "--migrate-config" to migrate config between different versions. This should activate automatically if an old version is detected.
If no config file is detected, use default values and save at the end.
Adds ability to have the config file in the same directory as this script.
Adds the ability to specify a custom directory for the venv. The config file must be in the same directory as this script for that to work.
Made persistent versions of nohtmlparser and nosslverification
Fixes the option for no ssl verification. This broke because I was using a legacy interface (urllib.request.urlretrieve) instead of the recommended one (urllib.request.urlopen).
Rename some commandline arguements to be more reasonable.
Use new config file for version check and fix positioning of check.