Implement game tweak intballseffectbellysize
The weight calculations are taken straight from the balls weight calculation but the size calculations are a bit arbitrary. The multiply by 0.9 from the size calculation comes from the approximate conversion from ballSize to inches (the unit the game uses) however the divide by 5 is more of a placeholder value until I figure out what magic the game uses for the actual calculation.