Add the original game's load failsafes when loading sol/nim files
By the looks of it, these were there to load save files from older versions of the game. I do not have any of those save files or an old version of the game so I can not test if the bag failsafe actually works (the others should work because the values are just missing and ".get" can use default values).