More optimisations
Remove the extra hideDiscard call in doReturn
Use sets in a lot more places where checking "if thing in (...)"
Reverted some more preemptive optimisations I did
Removed some checks of "== True" when checking booleans. I originally added these to most places because python is dynamically typed so I have no way of knowing what type an object will be but this can be removed for internal variables that aren't set to the output of a function since they are only modified by this program.