Switch back to text display stuff similar to the original game
The reason I tried to minimise the use of outputMainText in the first place was because it was extremely slow. Doing that caused way too many problems so I've switched everything to a similar system with two major differences, 1) Use a modified io.StringIO object to store the text while being used, and 2) Only update the text display at the end instead of updating it every time the stored text changes.

Also fixes a typo in the wiki changes page.