Optimise some of the bag button logic
Add a default value to the which arguement in showPage so it doesn't have to be specified when not needed.
Specify text when showing buttons/amounts instead of after.
Don't check if the item exists in bsRefresh because it always will.
Pass the buttonDict to showButtonsBag instead of the name array in choiceListButtons to remove a couple of extra steps. doButtonChoices no longer needs to be called after this which also allows removing choiceListBlanks from most places because it isn't needed.
Add bagDisableEmpty for where the previous change does not work.